updates based on review with exception for the performance issue and rebuilding the tree on data changes
